error message on invalid log in instead of crash
